Start building a list of commands to add in README
Add a few sample commands to the list that we might add that we have in
the README.  A build command would probably want to have
openshift/imagebuilder handle the file parsing, with us providing an
Executor for one of its Builder objects.  A run command would use runc,
or another OCI-compatible runtime, or maybe bubblewrap.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>