Add ImportBuilderFromImage()
Add a function for returning a Builder initialized from an image's
metadata, but which isn't tied to a container, so it can't be saved.

This refactored the existing importBuilder() implementation to split the
bits which locate the image based on the information it's given about a
container from the bits which build the structure using information read
from the image.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #109
Approved by: rhatdan