Show Config/Manifest as JSON string in inspect when format is not set
In https://github.com/projectatomic/buildah/issues/363:
Config and Manifest are displayed as a string representing JSON
only when "--format" is set. However, this is necessary too if
"--format" is not set.

This patch fixes this by using the same technique above
to make the representation of Manifest and Config consistent.

Signed-off-by: Boaz Shuster <ripcurld.github@gmail.com>

Closes: #522
Approved by: rhatdan