Manage "Run" containers more closely
Instead of using the runtime's "run" command to have it handle the
entire life cycle of a process when we need to launch one, do it
ourselves, and handle passing the data between our stdio and the
container's.

This will make it possible for us to set up networking using CNI between
the "create" and "start" phases, and head off permissions problems when
the process in the container can't read or write to the invoking user's
terminal or stdio.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #708
Approved by: rhatdan