buildah images and podman images are listing different sizes
After reviewing how the sizes are shown.  It looks like podman was
using docker/units to display size.  Its size is based off of multiples
of 1000, which according to Wikipedia is correct for image sizes.

https://en.wikipedia.org/wiki/Binary_prefix

Buildah was using multiples of 1024.  This patch switches buildah to use the same
multiple and to reduce its accuracy to two decimal points so that buildah
podman will return the same sizes.  This also will match the way that docker
reports image sizes.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>