Always set mappings when using user namespaces
If we're creating a user namespace, we always need to supply at least
one mapping for the UID and GID maps.  If we're not given any mappings,
map the ranges that are available to us, instead of assuming we can map
all possible values, in case we're already in a user namespace.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>