Fix buildah from docker.io/centos:7 pull
Currently buildah does not parse the docker.io syntax correcly.
In the example above it should parse this to docker.io/library/centos:7

The change in this patch will check to see if the user specified a transport
if not, then it will append the default transport "docker://" and containers/image
takes care of the rest.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>