Extend the amount of time travis_wait should wait
The travis_wait wrapper should be giving us 20 minutes by default, but
that's for the entire runtime of the command we're running under it.

While an individual 'make rpm' test can take more than 10 minutes, it
shouldn't take _much_ more, so increase the allowed total to three times
a reasonable number (say, 15), plus some time for all of the other tests.

When travis_wait runs out of time, it tries to kill the job, which we're
running under sudo anyway, so things can get weird.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>