runCollectOutput(): just read until the pipes are closed on us
runCollectOutput() tries to read error messages that may have been
written to pipes by the container runtime.  Instead of setting them to
non-blocking and producing an error when we fail to read data, leave
them blocking so that we wait until the write end of a pipe is closed
before we give up on reading from it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>