Do not add EmptyLayer entries to DiffID lists in the manifest
When containers/image/storage copies a schema1 image to containers/storage,
it creates a physical layer for all incoming layers, including those marked
as Throwaway.

When committing an image, we construct the list of layers by walking the
parent relationship, including the Throwaway / EmptyLayer layers.  Those should
not appear in the config's RootFS.DiffID, though, and a consistency check
was flagging this as an inconsistency.

So, check for the EmptyLayer flag in the History array, and do not add such
layers to DiffID.

This adds an assertion that the History length matches the list of layers
through the parent relationship; that was already assumed by the DiffID
consistency check, we are only moving the check forward.

A plausible alternative to this might be to not create physical layers for
the Throwaway/EmptyLayer values at all.  I don't know which one actually
makes more sense; this fix is at least localized and short, not having the
physical layers would require changes both in containers/image/storage and
here (to modify the aforementioned History length assumption).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>