Run(): tweak terminal mode handling
Default to using a controlling terminal if all three stdio descriptors
are terminals, not just stdout.

Don't try to set stdin to raw mode while running a container if it's not
a terminal, in which case it doesn't support terminal modes.

Don't try to read the window size of stdin if it's not a terminal, in
which case it doesn't have a window size.  Provide a way to explicitly
set it for those cases.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>