Use c/image/manifest.GuessMIMEType instead of a custom heuristic
Notably this will use the manifest mediaType field, instead of the
config docker_version field, to detect Docker v2s2 images; again,
docker_version is cleared by buildah itself, so it can't be relied
to decide on the image format.

Anyway, centralizing the tricky heuristic and using a single one
everywhere seems preferable for consistency.

This also allows only searching for OCI annotations on known OCI
images, and that in turn allows us not to silently ignore errors
when parsing OCI manifests.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>