Don't throw away the manifest MIME type and guess again
When loading the manifest using a types.Image.Manifest,
we also get a manifest MIME type; don't just throw it away
and then guess as if the data was never available, just use it
directly (now that the loading is close enough for the throwing
away to be noticeable).

[Ultimately, this loads the manifest from a c/image/storage
backend, which calls manifest.GuessMIMEType anyway, so this
does not change behavior right now, but if the c/image/storage
backend ever changed to record MIME types, it would make a
difference.  At the very least, we now run the heuristic once
instead of twice.]

Signed-off-by: Miloslav Trmač <mitr@redhat.com>