Use the c/image conversion code instead of converting configs manually
Now that we have a types.Image available in Builder.initConfig, we can
let it do the heavy work of converting configs to the v2s2 and OCI formats.

This requires us to cheat a bit with c/image/types.InformationOnly, but it
works out fine.

AFAICS this should have equivalent effect to the previous implementation,
except:
- Conversion from OCI will no longer crash if "Created" timestamps are missing.
- Conversion from v2s2 to OCI only uses the Config field, not ContainerConfig.
  Considering the Docker deamon only ever uses Config for populating the
  configuration of newly created containers in (docker create), this seems
  good enough.  But, to be honest, I'm not quite clear on the purpose of the
  Config/ContainerConfig distinction in v2s2.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>