Separate stdio buffering from writing
When relaying stdio data to and from a container, separate the "reading
and buffering from a read descriptor" step from the "relaying to a write
descriptor" step, so that we can try to flush buffers that have data in
them even when there's no new data to be read.

Treat EAGAIN as a recoverable error when writing, since we're now able
to come back and try again later.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>