Allow --userns-uid-map/--userns-gid-map to be global options
Allow --userns-uid-map and --userns-gid-map to be specified as global
options, which can be overridden by options specified to commands which
know them as non-global options.

This will mainly benefit tools that want to set ID mappings for buildah
to use, but which don't want to have to figure out whether or not
they're invoking it with a subcommand that recognizes these options.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>