Run(): try to resolve non-absolute paths using $PATH
Try to resolve commands which aren't given as absolute paths using the
$PATH environment variable and the mounted rootfs.  If we don't have a
configured $PATH, add one.

We can't resolve symbolic links with absolute values reliably without
using chroot(), so we just take it on faith that a link, or a non-link
with the execute bit set, will work.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #820
Approved by: rhatdan