Fix rmi to remove intermediate images associated with an image
buildah rmi now removes cached images associated with an image
when the image is removed only if the cached image is not required
by another image.
If an image is a parent and has a name then untag, otherwise return
an error stating the image can't be remoed as it has dependent
child images.

Signed-off-by: umohnani8 <umohnani@redhat.com>

Closes: #871
Approved by: rhatdan