Add and implement IsolationOCIRootless
Add an IsolationOCIRootless that runs the OCI runtime with its
--rootless flag, with network and UTS namespaces disabled, with IPC,
PID, and user namespaces forcibly enabled.  In this mode, we don't
attempt to set the container's hostname (because we don't have our own
UTS namespace), and we don't try to set any supplemental groups.  The
/sys directory is replaced with a bind mount of the host's /sys rather
than a fresh sysfs instance.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>