Eliminate the long-running ref and img variables in resolveImage
Inspecting the code flow, img seems to end up nil nil when we exhaust
all options and exit the loop, so img and ref don't need to be long-lived.

Should not change behavior, except possibly for a theoretical
case where store.Image(image) succeeds but !strings.HasPrefix(img.ID, image);
OTOH in that case, if it could happen, we would end up with a nil ref, so
that was very likely never intended in the first place.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>