Don't bother with --no-pivot for rootless isolation
When running outside of a container, --no-pivot isn't necessary, and
when running inside of a container, it's not enough to solve any of the
difficulties we're seeing there.  It may trigger an EPERM for unshare()
calls inside of the container that we launch, and we don't want that, so
drop it, for now at least.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>