Fix --layers ADD from url issue
When building with layers and using the ADD command to
add a file using a url, the build process would get stuck
in the second round when using the cache. This was happening
because the node traversal code was not moving on to the next
node when a url was detected hence causing an infinite loop.

Signed-off-by: umohnani8 <umohnani@redhat.com>