Make buildah-from error message clear when flags are after arg
The error message in buildah-from can be confusing when a
flag is specified after the argument (image name).

For example, running the following:
buildah from scratch -q

will return this error message:
too many arguments specified

however only one argument was specified "scratch".
the reason we get this error is because anything that
comes after "scratch" will be considered as an argument.

To fix this - check if every argument is a flag (a flag
starts with "-" e.g. "-q", "--ulimit"). If you find an
argument that is a flag return a much clearer
error message:

<flag-name-argument> should be set before the image name

otherwise, return the old error.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>

Closes: #968
Approved by: rhatdan