Use random string for getting a unique container name
Currently a buildah bud unique container name is found by incrementing
a counter and checking if that container exists. This causes a race
where that name may be claimed by another run by the time the
container is created.

This change replaces the incremented counter with a random string, so
that no duplicate name check is required.

Fixes #1004

Signed-off-by: Steve Baker <sbaker@redhat.com>