new: Return image-pulling errors from resolveImage
Instead of throwing out the upstream error message, save it and return
it if we can't come up with a more-specific suggestion.  This makes
debugging easier [1].

Using multierror gives us both pretty formatting (when we print this
for the user) and programmatic access (for any callers that need to
inspect the constituent errors).  This is our first direct dependency
on multierror, but we've been vendoring it for a while now because
opencontainers/runtime-tools uses it for config validation.

The error message for the Bats pull is now:

  error creating build container: 1 error occurred:

  * Error determining manifest MIME type for docker://non-existent-registry.com/alpine:latest: pinging docker registry returned: Get https://non-existent-registry.com/v2/: dial tcp: lookup non-existent-registry.com on 172.16.69.5:53: no such host

[1]: https://github.com/projectatomic/buildah/issues/849

Signed-off-by: W. Trevor King <wking@tremily.us>

Closes: #1014
Approved by: rhatdan