Add a blob cache implementation
Add a blob cache reference type which wraps other reference types.  When
used to create source and destination images, it returns wrappers for
source and destination images created by the wrapped references.

The cache reference must be created with a specified directory.

Blobs that are written to a cache image destination are stored as-is on
disk.  When reading a blob from a cache image source, if a blob with the
matching digest (and if specified, a matching size) can be found on
disk, read it directly instead of reading it from the real image source.

If compression or decompression is requested, the cache may return a
related blob instead of an exactly-matching blob.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>