pullImage(): when completing an image name, try docker://
If we're handed a name for a source image, and
alltransports.ParseImageName() doesn't like it as-is, we normally try
prefixing the Transport option and parsing it again.  If that option
isn't set, though, we return an error, when we should be trying the
default ("docker"), which is how NewBuilder() does things.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #1072
Approved by: rhatdan