Wire in a hidden --blob-cache option
Add a hidden --blob-cache option to bud, commit, from, pull, and push,
mainly so that we can exercise the APIs using integration tests.  It's
easy to accidentally get to a place where an image created while using
this option at various times won't be pushable, so for at least now, it
should stay hidden and undocumented.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #1149
Approved by: rhatdan