rootless: only discard network configuration names
When changing settings for rootless containers, only discard the part of
the networking configuration that specifies which networks we want to
configure, and preserve whether or not we want to use the host's network
namespace.

If we were told to create an empty network namespace or to just attach
to another namespace, go ahead and try to do that.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>