Add --disable-compression to 'build-using-dockerfile'
Add a --disable-compression flag to the 'build-using-dockerfile'
command, and set OCIAcceptUncompressedLayers and DirForceCompress in the
SystemContext during Commit() and Push() based on whether or not the
flag is set.

For other transports, we still end up going with the receiver's
preference, so update the man page to mention that this controls a
default.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>

Closes: #1149
Approved by: rhatdan