Fix cache issue with COPY/ADD --from
In multi-stage builds, it is possible to copy data from
a previous build. When using cached images, buildah wasn't
able to detect a change in the previous build and would proceed
to use cached images. Fix buildah bud to be keep track of any
files coming from a previous build and checking whether the cached
image is stll valid or not.
Also fix removal of intermediate build containers when building
from cache. We were not keeping track of those containers so they
weren't getting deleted at the end of the build process.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>