Update for API changes
NOTE: This is a very minimal update; blobinfocache.go should be
substantially reworked to benefit from the cache instead of decompressing
layers to get the DiffID values.

Also removes some HasBlob checks in the blobinfocache test suite which don't have
a direct equivalent in the new API, and makes the test suite independent
of host's signature policy.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>

Closes: #1218
Approved by: rhatdan