blobcache.bats: adjust explicit push tests
Remove error-prone parts of the explicit push tests that push to a
directory without using the blob-cache.  The proneness relates directly
to the fact that the layers of pulled images (i.e., the busybox base
image) are compressed with another gzip library the blobs in Buildah's
cache, which now use pgzip.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>