Add CreatedAtRaw date field for use with Format
Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>

This adds the ability to see a "raw" date format for the CreatedAt field rather than the pretty formatted variety.

Wit this in place the files will print thusly:

.CreatedAt: "Jan 8, 2019 09:22"
.CreatedAtRaw: "2019-01-08 09:22:52.330623532 -0500 EST"

This hopefully at least partially addresses #1266
Once agreed to, we might want to do a similar thing for buildah containers and then in similar commands in Podman.

Closes: #1269
Approved by: rhatdan